Rename CompareLoadoutsDrawer
