+++ src/app/loadout-builder/generated-sets/CompareLoadoutsDrawer.m.scss
